# frontend/generated/backend-openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "HTTPErrorResponse",
        "type": "object"
      },
...
      "ProjectIngestJob": {
        "properties": {
          "audio": {
            "title": "Audio",
            "type": "boolean"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "number"
          },
          "duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "fps": {
            "title": "Fps",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "projectName": {
            "title": "Projectname",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "resolution": {
            "title": "Resolution",
            "type": "string"
          },
          "video_path": {
            "title": "Video Path",
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectName",
          "video_path",
          "prompt",
          "model",
          "resolution",
          "duration",
          "fps",
          "audio",
          "createdAt"
        ],
        "title": "ProjectIngestJob",
        "type": "object"
      },
      "ProjectIngestListResponse": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/ProjectIngestJob"
            },
            "title": "Jobs",
            "type": "array"
          }
        },
        "required": [
          "jobs"
        ],
        "title": "ProjectIngestListResponse",
        "type": "object"
      },
...
      "StatusResponse": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "LTX-2 Video Generation Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/project-ingest": {
      "get": {
        "operationId": "route_list_project_ingest_api_project_ingest_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectIngestListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route List Project Ingest",
        "tags": [
          "project-ingest"
        ]
      }
    },
    "/api/project-ingest/{job_id}": {
      "delete": {
        "operationId": "route_delete_project_ingest_api_project_ingest__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Route Delete Project Ingest",
        "tags": [
          "project-ingest"
        ]
      }
    },
...
  }
}